// src/types/UserOperation.ts
export interface UserOperation {
    sender: string;
    nonce: number;
    chainId: number;
    initCode: string;
    callData: string;
    callGasLimit: number;
    verificationGasLimit: number;
    preVerificationGas: number;
    maxFeePerGas: number;
    maxPriorityFeePerGas: number;
    paymasterAndData: string;
    signature?: string; 
  }
  